Eliminate unnecessary casts.
Use static_cast.

+++ similar/main/cmd.cpp
@@ -128,7 +128,7 @@ void cmd_parse(char *input)
 	char buffer[CMD_MAX_LENGTH];
 	char *tokens[CMD_MAX_TOKENS];
 	int num_tokens;
-	int i, l;
+	uint_fast32_t i, l;
 
 	Assert(input != NULL);
 
@@ -137,7 +137,7 @@ void cmd_parse(char *input)
 	strncpy( buffer, input, sizeof(buffer) );
 
 	//printf("lead strip \"%s\"\n",buffer);
-	l = (int)strlen(buffer);
+	l = strlen(buffer);
 	/* If command is empty, give up */
 	if (l==0) return;
 	
@@ -147,7 +147,7 @@ void cmd_parse(char *input)
 	//printf("trail strip \"%s\"\n",buffer);
 	
 	/* Split into tokens */
-	l = (int)strlen(buffer);
+	l = strlen(buffer);
 	num_tokens = 1;
 	
 	tokens[0] = buffer;
@@ -297,9 +297,8 @@ const char *cmd_complete(char *input)
 {
 	cmd_t *ptr;
 	cmd_alias_t *aptr;
-	
-	int len = (int)strlen(input);
-	
+	uint_fast32_t len = strlen(input);
+
 	if (!len)
 		return NULL;
